--- docs/src/docs/components/select-single.mdx
@@ -1,132 +1,0 @@
----
-title: Select
-description: Select Component
----
-
-# Select
-
-필드를 눌러 드롭다운 메뉴, 모달, 바텀시트를 호출하여 항목을 선택할 수 있습니다.
-
-## 활용
-
-- 여러 옵션 중 하나의 값을 선택할 수 있어요.
-- leftContent에 요소를 추가 할 때는 `SelectContent`로 감싸서 사용합니다.
-
-<Demo code={`import { SelectSingle, Option, OptionGroup, ListText, Label, FlexBox } from '@wanteddev/wds';
-
-const options = [
-  {
-    label: '옵션1',
-    value: 'option1',
-    group: '그룹1'
-  },
-  {
-    label: '옵션2',
-    value: 'option2',
-    group: '그룹1'
-  },
-  {
-    label: '옵션3',
-    value: 'option3',
-    group: '그룹1'
-  },
-  {
-    label: '옵션4',
-    value: 'option4',
-    group: '그룹2'
-  },
-  {
-    label: '옵션5',
-    value: 'option5',
-    group: '그룹2'
-  },
-  {
-    label: '옵션6',
-    value: 'option6',
-    group: '그룹2'
-  },
-  {
-    label: '옵션7',
-    value: 'option7',
-    group: '그룹3'
-  },
-  {
-    label: '옵션8',
-    value: 'option8',
-    group: '그룹3'
-  },
-  {
-    label: '옵션9',
-    value: 'option9',
-    group: '그룹3'
-  }
-];
-
-const Demo = () => {
-  const groups = options.map(option => option?.group);
-
-  return (
-    <FlexBox flexDirection="column" gap="8px">
-      <Label>Label</Label>
-      <SelectSingle defaultValue="option1" placeholder="선택해주세요.">
-      {options.map(option => (
-        <Option key={option.value} value={option.value}>
-       {option.label}
-        </Option>
-      ))}
-      </SelectSingle>
-    </FlexBox>
-  )
-}
-
-export default Demo;`}
-/>
-
-## Components
-
-아래와 같이 조합하여 사용합니다.
-
-```tsx
-<Select>
-  <OptionGroup>
-    <Option>
-      옵션1
-    </Option>
-  </OptionGroup>
-  <OptionGroup>
-    <Option>
-      옵션2
-    </Option>
-  </OptionGroup>
-</Select>
-```
-
-### Select
-
-<PropsTable component="SelectSingle" />
-
-### SelectContent
-
-`TextInputContent`와 동일한 props를 사용합니다.
-
-<PropsTable component="SelectContent" />
-
-### OptionGroup
-
-`MenuGroup`과 동일한 props를 사용합니다.
-
-<PropsTable component="OptionGroup" />
-
-### Option
-
-<PropsTable component="Option" />
-
-## Examples
-
-### Radio
-
-### Invalid
-
-### Disabled
-
-### LeftContent
