+++ packages/wds/src/components/select/constants.ts
@@ -1,4 +1,4 @@
-export const SELECT_SINGLE_NAME = 'SelectSingle';
+export const SELECT_NAME = 'Select';
 export const OPTION_GROUP_NAME = 'OptionGroup';
 export const OPTION_NAME = 'Option';
 export const SELECT_CONTENT_NAME = 'SelectContent';
